% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alFeatureGen.R
\name{alFeatureGen}
\alias{alFeatureGen}
\title{Andy Limber's Kaggle feature generation}
\usage{
alFeatureGen(data)
}
\arguments{
\item{data}{A dataframe containing the data for the
popularity classifications that will be modified. Can be train or test data.
no modifications are needed.}
}
\value{
The modified data with added columns. Currently adds a title length field and
year and month field.
}
\description{
Generates features from the original training and test data to be used in the classifier.
If to be used with alKK - must be called on both train and test data
}

