% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alRF.R
\name{alRF}
\alias{alRF}
\title{Andy Limber's Random Forest}
\usage{
alRF(train, test, NT = 1000, MT = 12, NS = 25, vars = NULL,
  seed = 123, Xtest = FALSE, CSV = TRUE)
}
\arguments{
\item{train}{A dataframe containing the training data for the
popularity classifications.}

\item{test}{A dataframe containing the data to be used to predict labels.
If a column of NAs is not supplied for the "popularity", one will be added.}

\item{NT}{Integer. The number of trees used by the forest (ntree).}

\item{MT}{Integer. The mtry parameter for randomForest.}

\item{NS}{Integer. The nodesize parameter.}

\item{vars}{Vector of column indices for the features to be used in the random forest.
Defaults to NULL and all features are used.}

\item{seed}{The seed to be used.}

\item{Xtest}{A logical that indicates whether you want to perform some
cross validation. Default is FALSE. If TRUE, will print an accuracy value.
Must supply values for popularity. Will over-ride other returned values.}

\item{CSV}{A logical that indicates whether a CSV of predictions should be
saved. Default is TRUE.}
}
\value{
Returns the predicted labels, probabilities and IDs, as well 
as a CSV of predictions if CSV is set to TRUE.
If Xtest is set to true, will return an accuracy.
}
\description{
Train a random forest classifier on training set and predict for test set. 
Can be done on its own for testing, but also will be run as part of 
the alKK function for final classification.
}

