% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alXval.R
\name{alXvalidate}
\alias{alXvalidate}
\title{Andy Limber's X validation}
\usage{
alXvalidate(n = 5, m = 9000, data, NT = 100, MT = 12, NS = 25,
  vars = NA, seed = 123)
}
\arguments{
\item{n}{The number of validations done.}

\item{m}{The size of the validation sample.}

\item{data}{The data on which Xvalidation will be done}

\item{NT}{Integer. The number of trees used by the forest (ntree).}

\item{MT}{Integer. The mtry parameter for randomForest.}

\item{NS}{Integer. The nodesize parameter.}

\item{vars}{Vector of column indices for the features to be used in the random forest.
Defaults to NA and all features are used.}

\item{seed}{The seed to be used.}
}
\value{
Vector of out of sample accuracy
}
\description{
Perform a cross validation on some set of data by dividing it into random samples, then 
using the alKK function on the training and test set created.
}

