% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alFeatureGen.R
\name{alFeatureGen}
\alias{alFeatureGen}
\title{Andy Limber's Kaggle feature generation}
\usage{
alFeatureGen(train, test, Xtest = FALSE)
}
\arguments{
\item{train}{A dataframe containing the training data for the
popularity classifications.}

\item{test}{A dataframe containing the data to be used to predict labels.
If a column of NAs is not supplied for the "popularity", one will be added.}

\item{Xtest}{A logical that indicates whether you want to perform some
cross validation at a later stage. Default is FALSE and a popularity column is added.
Should be set to TRUE if popularity column is to be unmodified, but non-NA values
must be provided.}
}
\value{
A list containing the new "test" and "train" dataframes with added features.
}
\description{
Generates features from the original training and test data to be used in the classifier.
}

