% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alKK.R
\name{alKK}
\alias{alKK}
\title{Andy Limber's Kaggle Klassifier}
\usage{
alKK(train, test, NT = 1000, MT = 12, NS = 25, imp = FALSE,
  seed = 123, Xtest = FALSE, CSV = TRUE)
}
\arguments{
\item{train}{A dataframe containing the training data for the
popularity classifications.}

\item{test}{A dataframe containing the data to be used to predict labels.
If a column of NAs is not supplied for the "popularity", one will be added.}

\item{NT}{Integer. The number of trees used by the forest (ntree).}

\item{MT}{Integer. The mtry parameter for randomForest.}

\item{NS}{Integer. The nodesize parameter.}

\item{imp}{Logical. Whether or not the importance should be calculated (default is false).}

\item{seed}{The seed to be used.}

\item{Xtest}{A logical that indicates whether you want to perform some
cross validation. Default is FALSE. If TRUE, will print an accuracy value.
Must supply values for popularity.}

\item{CSV}{A logical that indicates whether a CSV of predictions should be
saved. Default is TRUE.}
}
\value{
The random forest model as well as a CSV of predictions
if CSV is set to TRUE.
}
\description{
Assign popularity values to news articles for submission to Kaggle
}

